Created settings [create, read, update, delete], fix all request bad if status = 405, added permissions to user of company

+++ database/migrations/2019_07_06_151257_create_products_table.php
@@ -27,7 +27,7 @@ class CreateProductsTable extends Migration
             $table->integer('product_count');
             $table->boolean('execute')->default(0);
             $table->boolean('display')->default(1);
-            $table->integer('count_rates')->default(0);
+            $table->tinyInteger('count_rates')->default(0);
             $table->unsignedInteger('type_id');
             $table->unsignedInteger('user_id');
             $table->unsignedInteger('company_id');
